updates for MongoDB integration

+++ org.oobium.framework.tests/src/org/oobium/persist/http/models/CommentModel.java
@@ -68,14 +68,14 @@ public abstract class CommentModel extends Model {
 	/**
 	 * Get the count of all instances of Comment
 	*/
-	public static int count() throws Exception {
+	public static long count() throws Exception {
 		return getPersistService().count(Comment.class);
 	}
 
 	/**
 	 * Get the count of all instances of Comment using the given sql query and values.
 	*/
-	public static int count(String sql, Object...values) throws Exception {
+	public static long count(String sql, Object...values) throws Exception {
 		return getPersistService().count(Comment.class, sql, values);
 	}
